pyproject.toml
[project]
name = "bithumb"
version = "0.1.0"
description = "빗썸 거래소 자동 거래 프로그램"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pycopy>=2025.7.54",
    "pyinstaller>=6.16.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
